New IMAP ACL
Secure authentication based on the ``imaplib`` module.

Validating users against a modern IMAP4rev1 server that awaits STARTTLS
on port 143. Legacy SSL (often on legacy port 993) is deprecated and
thus unsupported. STARTTLS is enforced except if host is ``localhost``
as passwords are sent in PLAIN.

Python 3.2 or newer is required for TLS.