Implementing group restrictions
This patch add an new optional option 'group' in rights configuration file.
If specified, group membership of the user is verified before access is granted.
This need an auth backend implementing group membership. If not, access is
denied (security first). This patch implements it in PAM backend ;
implementation in LDAP backend is planned.