Fix etag generation
When PUTting an item, Radicale will:

* Store the text byte-for-byte
* Hash the text and return the etag

However, when fetching the item, Radicale will:

* Parse the text with the vobject library
* Serialize the text for output
* Derive the etag from the serialized text

When PUTting an item with UNIX-style lineendings, Radicale will return
the hash for the item with UNIX-style lineendings. However, when
fetching items, the UNIX-style lineendings are repaired and Radicale
returns a different etag.

A good idea would be to spare the vobject serialization/deserialization
roundtrip as it doesn't seem necessary, but this bug can also be fixed
by making the `get_etag` function indifferent towards various
lineendings.