Reliable slashless base URL detection.
This block of code is meant to translate
http://whatever.com/path/to/radicale into
http://whatever.com/path/to/radicale/.web
but since it needs to follow URL joining rules,
it returns "Location: radicale/.web",
repeating part of the URL back to the client.

But this mishandles http://whatever.com/path/to/radicale/,
since it detects this situation based on PATH_INFO,
assuming that an empty PATH_INFO means no trailing /,
This is (at least) not true for OpenBSD's httpd,
and maybe others.
Switch to looking at the unprocessed REQUEST_URI,
which reflects what URI the client thinks they're
accessing, and therefore properly combines with
the redirect we give them.