hook: gracefully ignore non functional hooks and fall back to none
In case a hook fails to load for some reason, fall back to the default
hook "none" and treat errors as warnings in the log.

This will gracefully ignore typos in hook names without crashing the
server, and it will also allow configuration of "rabbitmq" where i.e.
"pika" is missing.

Closes: #1490
Signed-off-by: Henning Schild <henning@hennsch.de>